requirements.txt
# Framework RAG
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.2.0

# Base vectorielle
chromadb>=0.5.0

# Embeddings
sentence-transformers>=3.0.0
langchain-huggingface>=0.1.0
faiss-cpu>=1.13.0  # For FAISS vector store

# LLM Providers
langchain-groq>=0.2.0  # Pour Groq (gratuit, Llama 3)

# PDF Processing
pdfplumber>=0.11.0  # For extracting text from PDF forms

# API Backend
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilitaires
python-dotenv>=1.0.0
jq>=1.8.0
pandas>=2.0.0  # Pour l'analyse des données CSV

# Tests
pytest>=8.0.0
httpx>=0.27.0
